Coloca as posicoes X e Y de PosicaoDTO como inteiros
Os atributos x e y da classe PosicaoDTO estavam marcados como floats,
sendo que representam, respectivamente, uma linha e coluna do grid.

+++ sketch/PosicaoDTO.pde
@@ -1,8 +1,8 @@
 public class PosicaoDTO {
-  public float x;
-  public float y;
+  public int x;
+  public int y;
   
-  public PosicaoDTO(float x, float y) {
+  public PosicaoDTO(int x, int y) {
     this.x = x;
     this.y = y;
   }
